Split out script annotation utils
